interview_backend/requirements.txt
absl-py==2.3.0
accelerate==1.7.0
addict==2.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asgiref==3.8.1
astunparse==1.6.3
attrs==25.3.0
audioread==3.0.1
bcrypt==3.2.2
beautifulsoup4==4.13.4
blinker==1.9.0
blis==1.2.1
Brotli==1.1.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudpathlib==0.21.1
colorama==0.4.6
confection==0.1.5
contourpy==1.3.2
crcmod==1.7
cryptography==43.0.0
cycler==0.12.1
cymem==2.0.11
datasets==2.21.0
decorator==5.2.1
deepface==0.0.93
deprecation==2.1.0
dill==0.3.6
ecdsa==0.19.1
editdistance==0.8.1
eyeD3==0.9.8
fastapi==0.115.12
fastapi-cli==0.0.7
ffmpeg-python==0.2.0
ffmpy==0.6.0
filelock==3.18.0
filetype==1.2.0
fire==0.7.0
flask-cors==6.0.1
flatbuffers==25.2.10
fonttools==4.58.2
frozenlist==1.7.0
fsspec==2023.6.0
funasr==1.2.6
future==1.0.0
gast==0.6.0
gdown==5.2.0
gevent==25.5.1
gradio==5.36.2
gradio_client==1.10.4
greenlet==3.2.3
groovy==0.1.2
grpcio==1.73.1
grpcio-status==1.62.3
gunicorn==23.0.0
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface==0.0.1
huggingface-hub==0.32.4
hydra-core==1.3.2
idna==3.10
intel-openmp==2021.4.0
itsdangerous==2.2.0
jaconv==0.4.0
jamo==0.4.1
jax==0.6.1
jaxlib==0.6.1
jieba==0.42.1
Jinja2==3.1.6
jmespath==0.10.0
joblib==1.5.1
jose==1.0.0
jsonpatch==1.33
jsonpointer==3.0.0
kaldiio==2.18.1
keras==3.10.0
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
libclang==18.1.1
librosa==0.11.0
llvmlite==0.44.0
lxml==5.4.0
lz4==4.4.4
marisa-trie==1.2.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mediapipe==0.10.21
mkl==2021.4.0
ml_dtypes==0.5.1
modelscope==1.28.0
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.1
mtcnn==1.0.0
multidict==6.6.0
multiprocess==0.70.14
murmurhash==1.0.13
mysql-connector-python==9.3.0
mysqlclient==2.2.7
namex==0.1.0
nest-asyncio==1.6.0
networkx==3.5
numba==0.61.2
numpy==1.26.4
omegaconf==2.3.0
openai-whisper==20240930
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.16.0
orjson==3.10.18
oss2==2.19.1
packaging==25.0
pandas==2.3.0
passlib==1.7.4
pillow==11.2.1
platformdirs==4.3.8
pooch==1.8.2
preshed==3.0.10
propcache==0.3.2
proto-plus==1.26.1
protobuf==4.25.3
psutil==7.0.0
pyarrow==20.0.0
pyarrow-hotfix==0.7
pyasn1==0.4.8
pyasn1_modules==0.4.1
pyAudioAnalysis==0.3.14
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.4
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.1
PyInputPlus==0.2.12
PyMySQL==1.1.1
pynndescent==0.5.13
pyparsing==3.2.3
PyPDF2==3.0.1
PySimpleValidate==0.2.12
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
pytorch-wpe==0.0.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
reportlab==4.4.2
requests==2.32.3
responses==0.18.0
retina-face==0.0.17
rich==14.0.0
rich-toolkit==0.14.4
rsa==4.9.1
ruff==0.12.3
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
semantic-version==2.10.0
sentence-transformers==4.1.0
sentencepiece==0.2.0
setuptools==80.9.0
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.2
soundfile==0.13.1
soupsieve==2.7
soxr==0.5.0.post1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy_pkuseg==1.0.0
spark-ai-python==0.4.5
srsly==2.5.1
starlette==0.46.2
stdiomask==0.0.6
sympy==1.14.0
tbb==2021.13.1
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
tensorflow==2.19.0
termcolor==3.1.0
tf_keras==2.19.0
thinc==8.3.4
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
tomlkit==0.13.3
torch==2.3.0
torch-complex==0.4.4
torchaudio==2.3.0
tqdm==4.67.1
transformers==4.52.4
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
umap-learn==0.5.9.post2
urllib3==2.4.0
uvicorn==0.34.3
wasabi==1.1.3
watchfiles==1.0.5
weasel==0.4.1
websocket==0.2.1
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.1.3
wheel==0.45.1
whisper==1.0.0
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.1
zh_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/zh_core_web_sm-3.8.0/zh_core_web_sm-3.8.0-py3-none-any.whl#sha256=7de3bd267176b9b2a8defb6997c1cd296da16c57b5e712f72ea44a51755421c8
zope.event==5.0
zope.interface==7.2
